src/requirements.txt
fastapi
uvicorn
starlette
pydantic
pydantic-settings
langgraph
langchain-core
langchain-openai
PyGithub
langchain-mcp-adapters
mcp
